P2/requirements.txt
django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
django-phone-field==1.8.1